+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,3 +1,5 @@
+package blackjack.domain
+
 class Card(
     val number: CardNumber,
     val shape: Shape
