+++ src/main/kotlin/blackjack/domain/DrawResult.kt
@@ -1,3 +1,5 @@
+package blackjack.domain
+
 sealed class DrawResult {
 
     object Success : DrawResult()
