+++ src/main/kotlin/blackjack/domain/DrawState.kt
@@ -1,3 +1,5 @@
+package blackjack.domain
+
 enum class DrawState {
     POSSIBLE, IMPOSSIBLE
 }
