+++ src/main/kotlin/blackjack/domain/GameResult.kt
@@ -1,3 +1,5 @@
+package blackjack.domain
+
 enum class GameResult(val description: String) {
     WIN("승"),
     LOSE("패"),
