+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -1,10 +1,12 @@
+package blackjack.domain
+
 class Player(
     val name: String,
     val cards: Cards = Cards()
 ) {
 
     private fun isPossibleToDrawAdditionalCard(): DrawState {
-        if (cards.getMinimumCardsValue() > BLACK_JACK_SCORE) {
+        if (cards.getMinimumCardsValue() >= BLACK_JACK_SCORE) {
             return DrawState.IMPOSSIBLE
         }
