+++ src/main/kotlin/blackjack/domain/PlayerGameResult.kt
@@ -1,3 +1,5 @@
+package blackjack.domain
+
 data class PlayerGameResult(
     val playerName: String,
     val gameResult: GameResult
